elfields_viewer/requirements.txt
pyelftools>=0.30